wgpu: Don't recompile shaders for each Pipelines
